xodtemplate: Avoid a ton of unnecessary strcmp()s
I know this won't make much of a difference, but I noticed the code, and
I couldn't help myself:

We first have this one, massive if () that checks if the define is any
of the valid object types, and if it isn't, we print an error, and exit.
Then, we loop over the valid object types, and check which one of them
it is.

Unless I'm crazy-ignorant about preschool math, if you already handle
all the valid cases, and you're left with another case, that's your
invalid case right there, without any extra checking of anything.

Change-Id: Ic00d2580602a8a195150f611daab973c2e3fecd7
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>