buildsystem: Use non-recursive make
I'm not going to argue why here. Instead I'm going to assume everyone
(who would have an opinion on the matter) is familiar with Peter Miller's
classic paper `Recursive Make Considered Harmful'.

This patch also adds a bit more structure to the project in general -
beyond the call of duty, if you will - by properly separating the lib
and src trees and *preparing* for a saner approach to binary namespacing.

We'll want to actually do the namespacing at some point as well (like,
let shadownaemon/naemonstats/... all have their own subdirectory in the
source tree) but this patch has gotten enough of a "THIS CHANGES
EVERYTHING" vibe to it already.

All test-suite related stuff is now tracked in tests.mk in the root.

Change-Id: I4d47d56c7cce2198418b571936f820325416eb23
Signed-off-by: Anton Lofgren <alofgren@op5.com>