workers: Log when notification command failed
There are very few things that make users as annoyed as missing
notifications - after all, if nagios doesn't notify, what good is it?
- so we really need to tell the user when the notification failed.

Ideally, this would be done through a notification. As we just
determined notifications doesn't work, we'll just have to log it
instead - perhaps you could monitor the log file and send a different
notification that works based on that?

This fixes bug op5#9353.

Change-Id: I9a5e19cc59d202189bb40e7cf43330017a3dcf06
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>