Refactor notification-type-to-string conversion
This removes a bunch of boiler plate, which makes the code a bit easier
to read.

In addition, it adds some type safety to the notification types by
declaring an enum over their possible values, which lets us get rid of
some (buggy - sizeof doesn't work like that) bounds checking.

Change-Id: Ib0b91f8f9896640863792c4c2d567212ecbe65c2
Signed-off-by: Anton Lofgren <alofgren@op5.com>