iobroker: Implement support for buffering outgoing data
The iobroker was really just a nm_poll(), which abstracted away the
underlying socket polling mechanism. This is a very useful
abstraction, but it doesn't abstract away as much as it should - I
don't really care to read or write to sockets, I'd like it to just
hand me data, or allow me to hand it data, and it could read or write
just fine, without me having to think about EAGAIN in 500 darn
places.

This necessitates that it uses two bufferqueues (an incoming, and an
outgoing), so it can respond responsibly to sockets glitching.

This implements the outgoing queue.

Change-Id: Ie80d4c0e243a705bbf2012960fed13978bdc9dae
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>