worker: Use nsock_write_all instead of only a meek write
This makes sure the full response is written to the master, before we
go on doing new jobs.

Unlike the last time I applied this patch - 2a04d62 - this will not
lead to deadlocks, because the master writes new jobs to the worker
nonblockingly, so it should get around to reading our response pretty
soon. If it doesn't, well, we've got bigger problems than a mere
worker can solve.

Change-Id: I4c4d91c38971041459953bf52fdfd5af505816ba
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>