Don't relog suppression reasons on a default level
This gets overly noisy. Instead, keep track of which of them have
been logged, and make them go to the debug log when we get repeating
NSR:s. They're only really interesting the first time anyway.

Of course, if the reason changes at some point, it will still be logged
as usual.

This complements #9442 according to feedback.

Change-Id: I9b815eb5360928dfb8f3e4e9b72f0a7bd78b7915
Signed-off-by: Anton Lofgren <alofgren@op5.com>