objects: Kill off should_be_scheduled
The new event queue doesn't respect it - it schedules everything with
enabled checks and a check interval, and if you don't want it to,
disable one of those.

Because the last object change of this kind was reverted and a
reserved variable put in its place, I'm doing that from the beginning.

Change-Id: Ib9ea1b86c01988e0dfa1d42f57215dc8a384873f
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>