checks: Fix next check handling for unchecked checks
When a check hadn't been checked before, the next check macro would
say that the next check time is 0 + check_interval. So all unchecked
checks would be attempted to be scheduled for the 1970's (this, of
course, fails, because we nudge the check to now(), so nothing
actually looks suspicious in the scheduling queue).

Now, at startup, we give all checks a different next check time, so
after they have been checked, this case will not trigger.

Unless, that is, the check wasn't/couldn't be completed, in which case
we would retry it instantly. The check would, most likely, fail again
- it is, after all, the same check that failed the last time - leading
to frequent and hard-to explain infinite loops where naemon would
insist on using 100% CPU, without doing anything.

Let's push it one check interval into the future. Even if it will
still fail then, at least other checks will have a chance to run in
the meantime.

And if check intervals are 0, still do not schedule it any more
recently than 1 interval_length into the future. Yield, you
uncooperative event, yield!

While at it, remove the next_host_check and next_service_check
variables we used to have, since they're not used anymore.

Commit-message-poetry-by: Robin Sonefors <robin.sonefors@op5.com>
Signed-off-by: Andreas Ericsson <ae@op5.se>