xpddefault: Make perffile processing asynchronous
This is achieved by sending perfdata to a bufferqueue, instead of
(necessarily) writing it straight to disk. While the perffile
processing job is taking place, we simply don't have any open perfdata
files and resorts to memory-only storage. When the job returns, we
reopen the file and flush the in-memory perfdata to the re-opened
file.

In my mind, this removes the last use we had for NPCD - the jobs are
now run asynchronously already, we don't need a separate daemon to run
them.

We still write to spool file per-checkresult, instead of bulking up
all the writes and writing them per-processing-interval (the
bufferqueue would have issued one write() per result anyway, but the
linux IO subsystem might have noticed the pattern and done Something
Clever™), because any pipe users would most likely have noticed the
change.

Change-Id: Iae95be538b38039b625765b3df7276f1cf0c1cea
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>