commands: Kill off the command worker
The old(est) command implementation had performance problems on busy
systems, so a command worker was written.

However, while the command worker was written, an in-core
query handler was also written. The query handler also accepts
commands. Apparently, that one has sufficient performance.

The standalone process that processes commands also had the problem
that while it was really, really fast to accept commands, it had to go
through the same interface as, say, the queryhandler to actually
submit the commands into the core.

So, because the actual performance should be the same, regardless of
whether a worker accepts the commands and forwards them or they go
straight to the core, this commit removes the command worker. Now,
external commands are handled much more closely to how the query
handler is handled.

Change-Id: I21cc3dcc94ddaba2d92fad7466fc7668029276aa
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>