Add kvvec_fetch and fix binary safe compare
kvvec contains keys with given values, which is often used to model a
key-value store. It's really seldom used with duplicate keys, and a
common operation is to retrieve the value for a given key.

kvvec_fetch implements the functionality to lookup a key_value struct
given a key, and given that the kvvec is sorted, it's done in O(log N)
time (otherwise O(N)).

It would be great to use the kv_compare to do the lookup, but since
that also compared the values, but didn't do it binary safe (which
kvvec should handle), kv_compare needed to be split up to a function
comparing to data blocks of different length deterministic and in a
binary safe manner, which kv_compare can use easily.

Change-Id: I32e9d6389a211e474c80fd4254a8bd621580cb5c
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>