Make tests use static libnaemon
Using the dynamically linked version of libnaemon makes libtool create
a wrapper script for tests, to load the newly compiled .so correctly.
That makes sense, and works quite well when running the tests directly,
or through "make check".

The problem is when a check segfaults, or want to easily be debugged
through valgrind, a directly accessable binary without any need for
patching library paths is really helpful.

Since the unit tests want to test just simple methods, and not the
integration, using the static version is sufficent.

Change-Id: I05d4e2e5ae538cf590c5895342f4db3d044f194e
Signed-off-by: Max Sikstrom <max.sikstrom@op5.com>