Automagically get version from git when we build
It's silly to track anything else, and this way we'll also have
the exact commit in the version string when people report bugs
and whatnot (assuming they ship that ofcourse, but they should,
or we can ask for it).

It also means we can get rid of the ridiculous "update-version"
script, which was only used to update the version in itself.
That is, you ran it with an argument, and it updated itself to
say that version. The version that the program(s) spit out
comes from configure.ac, which we now force to run a small
scriptlet in order to figure out which version we're at.

In all fairness though, that scriptlet has to be updated when
a new release is made, but that's kinda simple and we'll now
only track the version in *one* place.

Signed-off-by: Andreas Ericsson <ae@op5.se>