events: Move the main loop to naemon.c
Events now just carries functionality to poll for events - analogous
to how iobroker behaves. This means that the responsibility for
handling signals can be moved out of the event library, and suddenly,
the event library doesn't really depend on much of anything anymore.

This isn't the best solution - *something* (that isn't naemon.c and
also isn't logging.c) should take responsibility for log rotation, and
the event library should provide a way to register callbacks for
signals for that something to use. When that happens, it doesn't
matter if the event library has a main loop or not. However, this is
still a useful step: it shrinks the event library considerably, to the
extent that it would probably be advantageous to use the event library
instead of the raw squeue in the workers, which would make the
codebase more consistent.

To remove even more external deps, the nagios_iobs definition is moved
to the event library as well - because it should be managed by the
event library, eventually.

Change-Id: Ib543e557915954a1594cec92d7d90c8a03091d5a
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>