objects: Move timeperiods to their own object
They don't have to be squished into one single big-ass file with
everything else, they can be taken apart.

Considering the amount of querying going on in all the other objects'
interfaces (is child of, is contact for, etc), I'm a bit the check for
whether a timeperiod is active wasn't part of this commit. Oh well,
just moving the object related code from objects.c is a big enough commit.

Change-Id: Ia42e28db20dcda5c503acab6e5b2cc0e1f682dde
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>