cache horizons: Add extra sanity check
A horizon should apply if we're between when the check supposedly ran,
and $horizon seconds after the check supposedly ran.

We used to let the event queue move all the timestamps whenever a
clock change was detected, which meant that we could always assume
that the check wasn't lying about the last check time, even on a DST
change. However, the new event queue that uses monotonic time doesn't
care about timestamps, so we need to ensure that, when the time moves
1 hour backwards, we don't stop running checks for 1 hour - without
this patch we would, because in that case,

 last_check > current_time => last_check + check_horizon > current_time

So now we check that current time is between the start and end of the
horizon span, instead of merely checking the end. There are minor edge
cases - a jump 30 seconds backwards with a 30 second horizon will
cause some checks to be cached for 60 seconds once, and if the clock
moves forwards, the cache will be ignored for 1 check
interval. Neither of these two edge cases really matter, though.

For my own sanity, I've edited all horizon checks to have the same
format - all of them now say
  last + length > current && last <= current
and this means we're currently inside the cache horizon. Some of the
conditions had the arguments swapped, and some checked if we were
outside instead. Since my attempts to avoid off-by-ones and mistakes
in my de Morgan transformations caused my head to explode and my heart
to tear up ("Heart?", I hear you say. "Don't you mean eyes?" No,
because I don't have any eyes, because my head just exploded,
dumbass), I had to "standardize" it.

Change-Id: Idfbacf79163c98ce5525e95088de0e7ab091d7f4
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>