events: Don't necessarily shut down because there are no events
There might not be *timed* events, but that doesn't mean there are no
events, period - there are IO events and signals as well, that could
become relevant.

This doesn't matter right now, but it's more generic, so it would let
us move the "wait for workers to connect" code to the event loop, and
do the rest of the naemon setup in a callback.  Of course, this would
require the event library to handle IO events, but that's not too far
off, I think.

Change-Id: I1a14434832968be6b0a0e4972ab7c6eed43d574c
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>