Make NSR logging configurable
This patch adds the enable_notification_suppression_reason_logging
configuration directive. Just rolls off the tongue, doesn't it? NSR
logging is enabled by default, since it's useful for anyone who's
not running BSM & Merlin together, and only a nuisance if you are.

This works around #9586. The real fix should be done in BSM, by which
time this directive should be retired.

Change-Id: Ie4c92f5314201d2c16c6d286bd07eade852e874e
Signed-off-by: Anton Lofgren <alofgren@op5.com>