workers: Free memory after use, not before
When iocache was replaced with bufferqueues, it meant that
nm_bufferqueue_unshift_to_delim returned a newly allocated string, as
opposed to iocache_get_delim, which returned a pointer to an internal
structure. That meant that handle_worker_result needed to free the
string it got returned from worker_ioc2msg (which is just a wrapper
around nm_bufferqueue_unshift_to_delim). I remembered to free the
memory, but I didn't realize that buf2kvvec_prealloc returned a
structure pointing into the string we just passed into it (and, in
turn, that the wpres pointed to the kvvec's elements, and thus also
straight at the buf string as well).

Now, I free it at the end of the loop. As far as I know, that's when
iocache started to overwrite the memory, so it sounds like a safe
place to free it.

Change-Id: I44892bc72328c67b1c90e0f281c7c70bb4de9192
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>