Add GLib as a dependency
We have a whole bunch of sub-par data structures, string handling,
event loops, module loading, ..., et cetera implemented either in-core
or in libnaemon. This patch doesn't change any of that, but it *does*
give us the tools to do so with little risk and a lot of advantages
which will let us, instead of focusing on repairing API's and low level
utilities, focus on implementing a better monitoring solution and adding
new fabulous features.

There is no need for a clean switch. We can keep using our existing
utilities/APIs at our leisure, as we see fit. But - and this is key - we
don't have to!

Change-Id: Ie74aeda901ee08d49bd55d9b3a88f134586a854b
Signed-off-by: Anton Lofgren <alofgren@op5.com>