xodtemplate: Split the registration process in half
This first stage registers the object, and the second registers the
properties between the objects. By doing it this way, the second stage
can always assume that pointers to relations will resolve.

With this in place, we can put all of the duplication of objects into
the first stage, all of the recombobulation into the second step, and
the result of the duplication and recombobulation could be inserted
into the final objects directly instead of being cached in the
xodtemplate objects.

All of this is still todo - the road towards a more sensible
xodtemplate is long - but it's a step.

Change-Id: I7fc6629feae37f9d3fa7d8bf699dcea42de351b2
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>