build: Put all the files in libnaemon.so
libnaemon isn't really stable or versioned anyway, and if it was, it
would be most appropriate to keep common functions to interact with
nagios - as opposed to merely generic data structures - in this
library, so external tools (like shadownaemon) could easily reuse not
just the hashtable, but, say, the config parsing as well.

This makes test writing for nebmods suddenly sensible - a test can
easily register the module and simply trigger the right broker events,
rather than having to mock all of naemon's internals every time.

Change-Id: I0ec61381f26727b13faaeaa481a954d957ae2b8d
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>