objects_hostgroup: Use more pointers, use less strings
If the caller has an object, asking it to retrieve the name string
just so objects_hostgroup can lookup the object we had to begin with
is inefficient. If the caller doesn't have an object, it's easy to
lookup. Hence, make the function accept the object instead of the
string.

No, I didn't forget to port over any validation that the object
exists: xodtemplate_expand_hosts already does that.

Change-Id: I0aace45ed36c44deb533658eea9e655b291d9e99
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>