objects: Take control of all submitted string variables
This way, memory owned by objects is allocated by objects. This means
that the destructor maps to the constructor in a predictable manner,
so a third party module or config parser can behave predictably.

This leads to more strdups, which is slower than less strdups, but if
performance is your goal, you should get rid of the recombobulation
nonsense, and just create the real objects earlier in the
configuration parsing process, the way 681f04a hints at.

Change-Id: If824aa744fcac8f37ca77340409d40ac31b9a32c
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>