service: Make it possible to remove a service group membership
Just like the addition function, this is symmetric, meaning that you
call it once, and it breaks both relations.

This must obviously be called by both services' and servicegroups'
destructor, as they would otherwise leave dangling pointers in the
other party's objects. This is just as awkward as it sounds, and would
be helped tremendously by getting rid of the bloody servicesmember
struct already.

Change-Id: If17d49543a685ff918d47130ce5091c01ac1de4e
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>