host: Make group relationships removable
The motivation for adding only these is that these are common two-way
relations, which means that the destructor can be called from either
direction, leading to broken data on the other end. Ideally, all
relations - even one-way relations - should have removal functions,
but as long as it doesn't result in memory corruption, it's not really
a priority for me right now.

Change-Id: Ifd3822a6e9b5f5b60674795f55f2a427d66b9370
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>