Remove pointless bookkeeping from hosts
Hosts used to keep track of their average check interval and their
combined hourly_value.

For every place where the aggregate hourly_value is expected (when
deciding whether to notify about a host, and when setting the
PROBLEMVALUE macro), the services' combined hourly
value is already added to the host's, so adding the services' hourly
value at construction time as well is an obvious bug - so the test is
changed to test that things work correctly. And there are places - the
HOSTVALUE macro - that actually expect the non-aggregated
hourly_value, so we actually can't precompute the total.

In addition, the average check interval for deciding how to fake the
hosts check interval is just nonsense - the host has a check interval,
use that!

Change-Id: If0540b56d8276135abbc88ffc4dedbfeac813650
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>