objects: Remove as much validation as possible from configuration.c
It's not that I hate configuration.c (though I do), it's just that
most of the errors can much more easily be checked by the object
itself, a lot of it already is and the extra-checking was therefore
totally redundant, and the so-called validation also happens to set up
some objects with pointers that they should have just had to begin with.

Incidentally, this means that commands are actually included in the
config that oconfsplit generates.

Change-Id: I6cfd7369d6b4c8931b8bba49ec2e369803852f97
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>