worker: Stop trying to destroy events that are already destroyed
There are only two cases where a worker needs to destroy a timeout
event: when the job failed to spawn, and when the job completed ahead
of time. However, kill_job and destroy_job runs both when the job
completed ahead of time, and when the job is being killed by a
timeout, so we need to move the destroy_event call up in the call
stack for the case where it's needed, so we don't double-run it.

Change-Id: Id52f90672760bb468667d4498d16be5be750d266
Signed-off-by: Robin Sonefors <robin.sonefors@op5.com>