worker: Actually stop chdir()'ing anywhere
A "git commit --amend <NO PATH HERE>" fail brought to you by
yours truly is fixed with this commit.

If the main process is run as a daemon, it has already
chdir()'ed to where it should be and if it hasn't, we really
want any coredumps produced to end up where we can find them.

Change-Id: I5716f1e0f415ec09119e30751bfdb81c3abb96cc
Signed-off-by: Andreas Ericsson <ae@op5.se>